requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
streamlit==1.32.2
ultralytics==8.1.29
